Included Example Folder
Moved and renamed "statMatrix Example.m" to be "Introduction.m" inside the \Examples folder. Also created two empty files that will contain examples of how to compile the statMatrix data and then analyze it. I also included a copy of my Ensemble drift code to demonstrate how the statMatrix could be used in combination with the included data management functions.