Update
Updated with new plot: PlotTrialEventByTrialTypePEH_SM which runs through and plots trial types (position or odor) based on groupings (correct/incorrect inseq/outseq).

Made some quality of life changes to PlotTrialEventPEH_SM but that file needs to be reworked to be a single (recursive) for loop like the new plot is. It's not as simple as pulling the middle and bottom for loops from the new figure and plopping them into the old one but it's not that far off honestly. It's that with some minor tweaks to make it all play nice.