qol changes
Modified the PEH bin size, 0.125 seems to be a nice width. Also tweaked some of the ID for things. Also made printYN a variable at the beginning of the file. Most of the variables at the start of the file will be set as eventual inputs once this code is functionalized/automated.