Cleaning/Updating
Added in additional lines to make sure that the Odor/Position IDs were removed when no spikes occurred during that trial (lines 66/67 & 73/74).

Also updated it so that it is now able to be run using the previous odor ID instead of upcoming odor ID (upcoming odor ID is a straw man comparison, there's information about upcoming odor based on previous odor, not de novo like position is(ish)).

Also added in smoothing for the F-ratios to hellp clarify things visually... will probably need to remove this when doing actual analyses.