moving rms code Glenn found online
Gives basically the same output as when compared to a sliding window RMS approach as I had been doing before. Was integrated into code as a matter of convenience... may be removed at some point in the future as it seems to give a slight time or phase lag compared to running things with a nan padded sliding window.