Update
Added in InSeq v OutSeq difference section.... modified the base a bit to make it ammenable to running on either raw correlation values or rank values. Still need to tweak the wordings elsewhere... and it could all be a bit cleaner... still don't know what to do with this stuff... need to think on it some more.