New PLX analysis
Making a new code to pull out the PLX timestamps necessary for aligning the behavior and neural data. The old code works fine for rats trained on the traditional contingencies, however it doesn't work properly for rats trained on the SAS contingencies. This code is meant to be more dynamic and capable of handling both.