Added logical condition for pokes>targetDuration w/o buffer errors
The previous logic tossed to else and worked because if the buffer wasn't violated during the trial period, it would be violated by the period between trials. However, when the final trial occurred without a buffer violation, this logic failed and relied on an empty value for indexing ipiBufferViolation. Adding an additional logical 'elseif' here specific to this instance.