OE Updates
Updated the dialogue questions at initiation to account for the OpenEphys rigs.

Renamed the embedded behavior/neural matrix creation functions and limited their use to the PLX specific stuff. Will probably need to revisit this or create a new function specifically for experiments done entirely on the OpenEphys rig involving unit recordings as they will have .plx spike data and .continuous LFP data.

Began the process of creating the behavior matrix function for OE data. I realize now that I will definitely need to pull the ssnData file to compare with the OE events to ensure I'm correctly identifying trial parameters similar to how I need to consult it (or the .txt output) for the data recorded on plexon.