Adding flexibility for LFP data organization
Goal is to update the epoch extraction code to accept various inputs through varargin to modify what information is being compiled from the data set.

For now the only change made to the code is the removal of the requirement that the LFP column ID have a T([0-9]*) prefix to allow for use with the multi-site recording data which has different prefix structures.