Bugfix for trial time bins and secondary fix for previous issue with trials w/out events
These two issues were related to the common issue where the first trial in a session would potentially not have an event type associated with it. The timebins as well as the code to fill in absent trials with nans were both assuming the first trial would contain the event in question (which in hindsight was always problematic, especially for rear reward alignments).