Modified Behavior Matrix for Multi-List Sessions
Quick patch of the creator code to account for sessions where multiple odor lists were used on the Plexon rig. This easy fix resulted in a reversal of the order for Position and Odor columns in the behavMatrix. This won't cause a problem if indexing is being done logically (which it should be).