Update SummarizePLXevents_SD.m
Modified a check to trial event times that was likely catching instances where the last trial was started but never completed.

Also changed the poke buffer catches from being errors to being warnings meaning whomever is running the code will need to review the files for potential sources of the errors.